+++ custom_components/eplucon/eplucon_api/DTO/DeviceDTO.py
@@ -1,7 +1,8 @@
 from dataclasses import dataclass
-from typing import Optional
+from typing import List, Optional
 from .RealtimeInfoDTO import RealtimeInfoDTO
 from .HeatLoadingDTO import HeatLoadingDTO
+from .ZoneDTO import ZoneDTO
 
 
 @dataclass
@@ -12,3 +13,4 @@ class DeviceDTO:
     type: str
     realtime_info: Optional[RealtimeInfoDTO] = None
     heatloading_status: Optional[HeatLoadingDTO] = None
+    zones: Optional[List[ZoneDTO]] = None